Sync up before 11/30 session

+++ src/lib/abs_move_utils.h
@@ -83,6 +83,6 @@ typedef enum
 	GRABBER,
 	ROGGER_SLIDE
 } e_motor_move;
-#define distance_to_angle_derees(X) (X*360/ANGLE_SENSOR_CIRCUMFERENCE)
+#define distance_to_angle_derees(X) ((float)(X*360/ANGLE_SENSOR_CIRCUMFERENCE))
 #define distance_to_encoder_derees(X) (X*360/DRIVE_WHEELS_CIRCUMFERENCE)
 #endif /* !ABS_TURN_UTILS */
